requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
